headlamp-plugin: Change upgrade command to add some dependencies
@storybook/storyshots was for some reason not being added as a top
level dependency. I could not debug why, so we have to pin these
versions in the plugin itself. This makes it work, but is not
optimal of course.

Lucky the upgrade command can automate this for people.
I'm hoping newer dependencies don't have this problem in the
future.

Made it a bit easier to specify which keys should be copied
from the template with a replaceKeys helper function.